refactor styles on cameo & euphony page; begin to refactor homepage before final design
