+++ gatsby-browser.js
@@ -1,5 +1,5 @@
 import 'semantic-ui-css/semantic.min.css'
-import './src/styles/global.css'
+import './src/shared-styles/global.css'
 
 import CustomLayout from './wrapPageElement'
