+++ src/shared-styles/styles.ts
@@ -1,0 +1,23 @@
+const colors = {
+    black: 'black',
+    white: '#F6F6F3',
+    grey: '#424242',
+    beige: '#9e7762',
+    purple: '#d8c3d8',
+    offWhite: '#e3dede',
+    sweetBrown: '#A63D40',
+    green: '#7DAA7B',
+    redOrange: '#DA532D'
+}
+
+const fonts = {
+    megrim: 'Megrim',
+    montserrat: 'Montserrat',
+    zeyada: 'Zeyada',
+    yesevaOne: 'Yeseva One'
+}
+
+export {
+    colors,
+    fonts
+}
