+++ wrapPageElement.js
@@ -1,8 +1,7 @@
-import React from 'react'
 import PageLayout from './src/components/PageLayout'
 
 const wrapPageElement = ({ element, props }) => {
-  return <PageLayout {...props}>{element}</PageLayout>
+    return <PageLayout {...props}>{element}</PageLayout>
 }
 
 export default wrapPageElement
